+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+Version 1.2.0 *(In Development)*
+--------------------------------
+
+ * Support injection on any object using an Activity as the view root.
+ * Support injection on views for their children.
+
+
 Version 1.1.1 *(2013-05-06)*
 ----------------------------